requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.3.1
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.12.0
django-pusher==0.1.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
docopt==0.6.2
drf-yasg==1.21.8
Faker==33.1.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Js2Py==0.74
msgpack==1.1.0
multidict==6.1.0
ndg-httpsclient==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
pipwin==0.5.2
propcache==0.2.1
psycopg2==2.9.10
pusher==3.3.2
pyasn1==0.6.1
pycparser==2.22
pyjsparser==2.7.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.1.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
wheel==0.45.1
whitenoise==6.8.2
yarl==1.18.3
